+++ src/core/Board.hpp
@@ -10,9 +10,9 @@ class Board {
    public:
     [[nodiscard]] const std::unordered_map<Coordinate, Chunk>& get() { return board_; }
 
-    [[nodiscard]] bool isOccupied(int x, int y) const;
+    [[nodiscard]] bool isOccupied(int16_t x, int16_t y) const;
 
-    void set(TileKind kind, int x, int y);
+    void set(TileKind kind, int16_t x, int16_t y);
 
     void print()const;
 };
